+++ app/Http/Controllers/ItemController.php
@@ -154,7 +154,12 @@ class ItemController extends Controller
      */
     public function destroy( $id)
     {
-        $item = Item::findOrFail($id);
+        $item = Item::find($id);
+        if(!$item){
+            return response()
+                   ->json("Resource Not Found", Response::HTTP_NOT_FOUND);
+     
+        }
         $item->delete();
         return response(null, Response::HTTP_NO_CONTENT);
     }
