+++ app/Http/Controllers/SubscriptionController.php
@@ -247,7 +247,12 @@ class SubscriptionController extends Controller
      */
     public function destroy($id)
     {
-        $subscription = Subscription::findOrFail($id);
+        $subscription = Subscription::find($id);
+        if(!$subscription){
+            return response()
+                   ->json("Resource Not Found", Response::HTTP_NOT_FOUND);
+     
+        }
         $subscription->delete();
         return response(null, Response::HTTP_NO_CONTENT);
     }
